made a kinematics class and pulled all the methods out of the PhysicsWorld class that were from the first two chapters of the openstax college physics. I am aiming to make the PhysicsWorld more of a hub class that would contain some static structs and constants that are shared by many of the classes. I have begun adding more of the chapters as classes  and linking them to the PhysicsWorld so all I will need to do is add methods to the classes as I learn them.
