+++ PhysicsWorld.h
@@ -10,7 +10,6 @@
 
 #include <cmath>
 #include <iostream>
-#include <vector>
 #include "Drag.h"
 #include "DynamicsAndForces.h"
 #include "Elasticity.h"
@@ -23,6 +22,7 @@
 #include "Friction.h"
 #include "Hearing.h"
 #include "Heat.h"
+#include "Kinematics.h"
 #include "Momentum.h"
 #include "PeriodicElements.h"
 #include "RotationalMotion.h"
@@ -36,7 +36,7 @@
 
 
 class PhysicsWorld;
-//____________________________________________________________________________
+//_____________________________________________________________________________
 //global constants, methods and structures
 //
 //
@@ -102,14 +102,7 @@ inline ld delta(const ld xi, const ld xf)
 {
 	return xf - xi;
 }
-static ld _val_;// = 0.0;
-static void show_val() { cout << "_val_: " << _val_ << endl; }
-static void show_val(const ld val) { cout << "val: " << val << endl; }
-static void show_val(const string label) { cout << label+": " << _val_ << endl; }
-static void setVal(const ld v)
-{
-	_val_ = v;
-}
+
 
 /**
  * @brief structure for conversion methods
@@ -146,7 +139,8 @@ static struct Conversions
 	}
 
 	/**
-	 * @brief Returns the revolutions in radians per second which is the angular velocity as well
+	 * @brief Returns the revolutions in radians per second which is the
+	 * angular velocity as well
 	 * @param revMin revolutions per minute
 	 * @returns revolutions in radians per second
 	 */
@@ -156,7 +150,8 @@ static struct Conversions
 	}
 
 	/**
-	 * @brief Returns the conversion from revolutions per radian second to revolutions per minute
+	 * @brief Returns the conversion from revolutions per radian second to
+	 * revolutions per minute
 	 * @param radSec is the rotation speed in radians per second
 	 * @returns the revolutions per minute
 	 */
@@ -177,7 +172,8 @@ static struct Conversions
 
 
 	/**
-	 * @brief Returns the conversion from meters per second to kilometers per hour
+	 * @brief Returns the conversion from meters per second to kilometers
+	 * per hour
 	 * @param mps is meters per second
 	 * @returns kilometers per hour
 	 */
@@ -186,7 +182,8 @@ static struct Conversions
 		return mps * 3.6;
 	}
 	/**
-	 * @brief Returns the conversion from kilometers per hour to meters per second
+	 * @brief Returns the conversion from kilometers per hour to meters per
+	 * second
 	 * @param kmh is kilometers per hour
 	 * @returns meters per second
 	 */
@@ -231,7 +228,8 @@ static struct Conversions
 	}
 	/**
 	 * @brief Returns the conversion from days to seconds for use in calculation
-	 * @param days can be expressed as a decimal, four and a half days would be 4.5
+	 * @param days can be expressed as a decimal, four and a half days would
+	 * be 4.5
 	 * @returns total seconds converted from days
 	 */
 	static ld days_to_seconds(const ld days = _val_)
@@ -416,28 +414,6 @@ private:
 
 public:
 
-	static std::vector<ld> vector_values;
-	static void setVector(vector<ld> v)
-	{
-		vector_values = v;
-	}
-
-
-	 /**
-	 * @brief static member function to print values of passed in vector
-	 *
-	 */
-	template <typename T, size_t size>
-	static void show_vector_values(T (&size_t) [size])
-	{
-		for (auto it : size_t)
-		{
-			std::cout << it << ", ";
-		}
-		std::cout << std::endl;
-	}
-	static void show_vector_values();
-
 	/**
 	 * @brief Template method to print out an array that is passed to it
 	 * @param array of values
@@ -453,6 +429,7 @@ public:
 
 	//PhysicsCommon  physics_common;
 	//PhysicsWorld* world;
+	Kinematics* kinematics;
 	Friction* friction;
 	Drag* drag;
 	Elasticity* elasticity;
@@ -488,6 +465,7 @@ public:
 	PhysicsWorld(const PhysicsWorld&); //copy constructor
 	PhysicsWorld& operator=(const PhysicsWorld&); //copy assignment operator
 	PhysicsWorld(PhysicsWorld&& o) noexcept :
+		kinematics(o.kinematics),
 		friction(o.friction),
 		drag(o.drag),
 		elasticity(o.elasticity),
@@ -512,358 +490,6 @@ public:
 		electric_current(o.electric_current),
 		periodic_elements(o.periodic_elements),
 		_ptr_(o._ptr_){} // move constructor
-
-
-	static ld return_val() { return _val_; }
-	static vector<ld> return_vector(PhysicsWorld &v) { return v.vector_values; }
-
-	//============================================================================
-	//chapter 2 formulas
-
-	/**
-	 * @brief Returns the displacement between two positions
-	 * @param startPos starting position
-	 * @param endPos ending position
-	 * @returns displacement.
-	 */
-	static ld displacement(const ld startPos,const ld endPos)
-	{ return endPos - startPos; }
-
-	/**
-	 * @brief Returns the displacement from knowing the velocity and time.
-	 * @param velocity in m/s
-	 * @param time in seconds
-	 * @returns displacement.
-	 */
-	static ld displacement_VxT(const ld velocity, const ld time)
-	{ return velocity * time; }
-
-	/**
-	 * method: displacement_accelerating_object(ld acceleration, ld time)
-	 * arguments: acceleration m/s , time in s
-	 * purpose: find the displacement of an accelerating object
-	 * returns:	ld, displacement
-	 */
-	static ld displacement_accelerating_object(const ld acceleration, const ld time)
-	{ return ((acceleration) * (time * time)) / 2; }
-
-	/**
-	 * method: displacement_accelerating_object_PV(ld acceleration, ld time)
-	 * arguments: velocity, acceleration m/s , position = 0m default, time in s
-	 * purpose: find the displacement of an accelerating object with a starting
-	 * 			position and an initial velocity along with acceleration and a time
-	 * returns:	ld, displacement
-	 */
-	static ld displacement_accelerating_object_PV(const ld velocity,
-												  const ld acceleration,
-												  const ld time,
-												  const ld pos = 0)
-	{ return pos + (velocity * time) + (acceleration * (time * time)) / 2; }
-
-	/**
-	 * @brief calculates the displacement using the kinematic formula
-	 * X = Vi*t+ 1/2*a*t^2
-	 * @param velocity in m/s
-	 * @param acceleration in m/s^2
-	 * @param time in s
-	 */
-	static ld displacement_using_kinematic(const ld velocity,
-										   const ld acceleration,
-										   const ld time)
-	{
-		return (velocity * time) + (acceleration * (time * time)) / 2;
-	}
-
-	/**
-	 * method: displacement_halting_VdA(ld velocity, ld acceleration)
-	 * arguments: x0 = starting position(0), Sv = starting velocity, fV = ending velocity(0) a = acceleration
-	 * purpose: calculate displacement for the distance to stop from moving object
-	 * returns: ld, final displacement
-	 */
-	static ld distance_VdA(const ld pos, const ld velocityStart, const ld velocityFinal, const ld acceleration)
-	{ return abs(((velocityFinal * velocityFinal) - (velocityStart * velocityStart)) / (2 * acceleration)) + pos; }
-
-	/**
-	 * method: velocity_vStart_plus_vEndD2(const ld velocityStart, const ld velocityEnd)
-	 * arguments: vStart = starting velocity m/s, vEnd = ending velocity m/s
-	 * purpose:	find the displacement when start and end velocity is know
-	 * returns: ld, displacement
-	 */
-	static ld velocity_vStart_plus_vEndD2(const ld velocityStart, const ld velocityEnd)
-	{ return (velocityStart + velocityEnd) / 2;	}
-
-	/**
-	 *@brief calculates the final velocity using the kinematic equation Vf = Vi + a*t
-	 *@param initialVelocity is the initial velocity, many times this is zero
-	 *@param acceleration is the acceleration of object
-	 *@param time is amount of time taken
-	 *@returns the final velocity
-	 */
-	static ld velocity_final_from_kinematic_time(const ld initialVelocity, const ld acceleration, const ld time)
-	{	return initialVelocity + acceleration * time; }
-
-	/**
-	 *@brief calculates the final velocity using the kinematic formula vf^2 = vi^2 + 2*a*d
-	 *@param initialVelocity m/s
-	 *@param acceleration m/s^2
-	 *@param displacement m
-	 *@returns the final velocity
-	 */
-	static ld velocity_final_kinematic_no_time(const ld initialVelocity, const ld acceleration , const ld displacement)
-	{ return sqrt(initialVelocity * initialVelocity + (2 * (acceleration * displacement))); }
-
-	/**
-	 * method: velocity_avg_DdT(ld displacement, ld time)   ss2.3 pg37
-	 * arguments: displacement = avg displacement in m, time = total time  in s
-	 * purpose: to calculate the average velocity using displacement divided by time
-	 * returns: ld, average velocity
-	 */
-	static ld velocity_avg_DdT(const ld displacement, const ld time)
-	{ return  displacement / time; }
-
-	/**
-	 * method: PhysicsWorld::velocity_falling_object_down(ld y, ld yf, ld v, ld a)
-	 * arguments: y0 = start position,  yf = final position v = velocity, a = acceleration
-	 * purpose: find the velocity of a falling object thrown downwards
-	 * returns: ld, velocity
-	 */
-	static ld velocity_kinematic_constant_a(const ld xY0, const ld xYf, const ld velocity, const ld acceleration)
-	{ return sqrt((velocity * velocity) + (2 * (acceleration * (xYf - xY0)))); }
-
-	/**
-	 * @brief calculates the final velocity using the kinematic formula Vf = Vi*t
-	 * @param vi initial velocity
-	 * @param t time
-	 * @returns the final velocity
-	 */
-	static ld velocityFinal_kinematic(const ld vi, const ld t)
-	{
-		return vi * t;
-	}
-
-	/**
-	 * method: acceleration_avg(ld vChange, ld tChange)   ss2.4  pg40
-	 * arguments: vChange = average change in velocity in m/s, tChange = change in time in seconds
-	 * purpose:	to find the average acceleration by dividing the avg velocity by the time
-	 * returns: ld, average acceleration
-	 */
-	static ld acceleration_avg(const ld vChange, const ld tChange)
-	{ return vChange / tChange; }
-
-	/**
-	 * method: acceleration_dispDtimeSqrd(ld displacement, ld time)
-	 * arguments: 1)displacement 2)time
-	 * purpose: this method can be used to calculate the acceleration of falling objects
-	 * returns: ld, acceleration of object
-	 */
-	static ld acceleration_dispDtimeSqrd(const ld displacement, const ld time)
-	{ return (2 * (displacement)) / (time * time); }
-
-	/**
-	 * method: acceleration_vStart_vEndDdisplacement(ld velocityStart, ld velocityEnd, ld displacement)const
-	 * arguments: velocity start, velocity end, total displacement
-	 * purpose: finds the acceleration using the starting and ending velocity and total displacement as known values
-	 * returns: ld, acceleration of object
-	 */
-	static ld acceleration_vStart_vEndDdisplacement(const ld velocityStart, const ld velocityEnd, const ld displacement)
-	{ return (velocityEnd * velocityEnd - velocityStart * velocityStart) / (2 * displacement); }
-
-	/**
-	 * method: speed_avg_DdT(ld total_distance, ld time)   ss2.3 pg37
-	 * arguments: displacement = avg displacement in m, time = total time  in s
-	 * purpose: to calculate the average velocity using displacement divided by time
-	 * returns: ld, average velocity
-	 */
-	static ld speed_avg_DdT(const ld total_distance, const ld time)
-	{ return abs(total_distance / time); }
-
-	/**
-	 * method: time_by_avgVdA(ld acceleration, ld velocityStart, ld velocityEnd)
-	 * arguments: a = acceleration, sV = start velocity(default 0), fV = final velocity
-	 * purpose: calculate time with know acceleration and final velocity end point
-	 * returns: ld, time
-	 */
-	static ld time_by_avgVdA(const ld acceleration, const ld velocityStart, const ld velocityEnd)
-	{ return (velocityEnd - velocityStart) / acceleration; }
-
-	/**
-	 * method: time_by_DisTdV(ld distance, ld velocity)
-	 * arguments: distance , average velocity
-	 * purpose: find the velocity of a falling object thrown downwards
-	 * returns: ld, velocity
-	 */
-	static ld time_by_DisTdV(const ld distance, const ld velocity)
-	{ return distance / velocity; }
-
-	/**
-	 * method: time_finalPos_acceleration(ld displacement, ld acceleration)
-	 * arguments: displacement of object, acceleration of object
-	 * purpose: find how long it takes an object with a known acceleration to travel a know distance
-	 * returns: ld, time to accelerate a distance
-	 */
-	static ld time_by_finalPos_acceleration(const ld displacement, const ld acceleration)
-	{ return sqrt((2 * displacement) / acceleration); }
-
-	/**
-	 * method: time(ld startTime, ld endTime)
-	 * arguments: y = startTime = beginning time in seconds, endTime = endTime in seconds
-	 * purpose: find the amount of time between two periods
-	 * returns: ld, difference in two times
-	 */
-	static ld time_difference(const ld startTime, const ld endTime)
-	{ return endTime - startTime; }
-
-	/**
-	 * method: time_kinematic_rearranged(ld velocity, ld y0, ld acceleration) const
-	 * arguments: velocity, y0 = position from 0, acceleration
-	 * purpose: find the amount of time between two periods
-	 * returns: ld, difference in two times
-	 */
-	static ld time_kinematic_rearranged(const ld velocity, const ld displacement, const ld acceleration)
-	{ return (-(velocity)-sqrt((velocity * velocity) - 2 * (acceleration) * (displacement))) / (acceleration); }
-
-	/**
-	 * method: slope_formula(ld y1, ld y0, ld x1, ld x0)
-	 * arguments: y1 , y0, x1, x0
-	 * purpose:	the general slope equation of (y1 - y0)/(x1 - x0)
-	 *  to find the slope of a line between two points
-	 * returns: ld, slope of line between two points
-	 */
-	static ld slope_formula(const ld y1, const ld y0, const ld x1, const ld x0)
-	{ return (y1 - y0) / (x1 - x0);	}
-
-	/**
-	 * method: rotation_speed_2PIxRdT(ld radius, ld rotations, ld time)
-	 * arguments: radius = length in m from center of rotation
-	 *	rotations = how many rotations in a time period
-	 *	time = time units
-	 * purpose: find the speed of a spinning object, such as fan blade
-	 * returns: average speed of spinning object
-	 */
-	static ld rotation_speed_2PIxRdT(const ld radius, const ld rotations, const ld time)
-	{ return (2.0 * _PI_ * radius) / (time / rotations); }
-
-	/**
-	 * method: rotation_avgVelocity_2PIxRdT_in_1_rotation(ld radius, ld time)
-	 * arguments: radius , time = time for one rotation
-	 * purpose: find average velocity of a spinning object
-	 * returns: ld, average velocity
-	 */
-	static ld rotation_avgVelocity_2PIxRdT_in_1_rotation(const ld radius, const ld time)
-	{ return (2.0 * _PI_ * radius) / time; }
-
-	/**
-	 * @brief Returns the conversion of given value divided by the acceleration of gravity on earth, 9.80 m/s^2
-	 * @param value the value to find the multiples of gravity of
-	 * returns: g's
-	 */
-	static ld conversion_multiple_of_gravity(const ld value)
-	{ return value / GA; }
-
-	/**
-	 * method: time_using_quadratic(ld a, ld b, ld c)
-	 * arguments: a1 = default to 1, b_velocity = constant velocity, c_displacement = total distance to travel
-	 * purpose: calculate the time of a merging object when velocity and displacement is know
-	 * returns: ld, total time
-	 */
-	std::vector<ld> time_using_quadratic(ld a1, ld b_velocity, ld c_displacement)const
-	{
-		vector_values[0] = (-b_velocity + sqrt((b_velocity * b_velocity) - 4 * a1 * c_displacement)) / (2 * a1);
-		vector_values[1] = (-b_velocity - sqrt((b_velocity * b_velocity) - 4 * a1 * c_displacement)) / (2 * a1);
-
-		return this->vector_values;
-	}
-
-	/**
-	 * method: pos_vel_falling_object_upDown(double p, double v, double a, double t)
-	 * arguments: p = position (0),  v = velocity, a = acceleration, t = time is s
-	 * purpose: this method will fill a vector with four pieces of data in order from the right to left it
-	 *			is: time, position, velocity, acceleration. use the print_vector_values() to see contents.
-	 * returns: ld, vector of the time, position, velocity, acceleration.
-	 */
-	std::vector<ld> pos_vel_falling_object_upDown(ld v, ld a, ld t, ld p = 0.0)const
-	{
-		this->vector_values[0] = t;
-		//solution for Position:
-		PhysicsWorld obj;
-
-		this->vector_values[1] = obj.displacement_accelerating_object_PV(v, a, t, p);
-		//Solution for Velocity:
-
-		_val_ = obj.velocity_final_from_kinematic_time(v, a, t);
-		this->vector_values[2] = _val_;
-
-		this->vector_values[3] = a;
-		//this->show_vector_values();
-
-		return obj.vector_values;
-	}
-
-	//=================================================================================
-	// chapter 3 formulas
-
-	/**
-	* method: projectile_range_level_ground(ld velocity, ld theta)
-	* arguments: velocity, theta (initial angle relative to the horizontal
-	* purpose:	calculates projectile range
-	* returns: ld, range of a projectile
-	*/
-	static ld range_projectile_level_ground(const ld velocity, const ld angleTheta)
-	{ return  ((velocity * velocity) * sin(2 * angleTheta * RADIAN) / (GA));	}
-
-	/**
-	 * @brief calculates the angle theta of a projectile
-	 */
-	static ld projectile_theta(const ld distance, const ld velocity)
-	{
-		return (asinh(distance * _Ga_) / (velocity * velocity)) / 2;
-	}
-
-
-
-	/**
-	 * @brief calculates the time a projectile with an initial velocity and angle
-	 *  take to reach the same level from which it was launched, forms a parabolic curve
-	 * (2* launchVelocity*sin(angleTheta))/(-GA);
-	 * @param launchVelocity is the initial velocity in m/s
-	 * @param angleTheta the angle of the launch
-	 * @returns ld, time units
-	 */
-	static ld time_for_projectile_to_reach_level(ld launchVelocity, ld angleTheta)
-	{
-		return (2 * launchVelocity * sin(angleTheta)) / (GA);
-	}
-
-	// sqrt((-2*(displacement))/-GA)
-	ld air_time_initial_velocity0_y0(ld displacement)const;
-
-	//  sqrt((-GA/(-2 * y0)))* displacement
-	ld velocity_initial_horizontal_component(ld y0, ld displacement)const;
-
-	//velocity_final_vertical_component(ld y0) const
-	ld velocity_vertical_component(ld y0, ld yf = 0)const;
-
-	// sqrt(velocityY * velocityY + velocityX * velocityX);
-	// atan(velocityY / velocityX) * DEGREE;
-	std::vector<ld> final_projectile_velocity_vector(ld velocityY, ld velocityX)const;
-
-	// sqrt(pow(horizontal_velocity_using_distance_angle_height(toGoal,height_at_goal, angle), 2) +
-	// pow(vertical_velocity_by_Xvelocity_with_angle(horizontal_velocity_using_distance_angle_height(toGoal, height_at_goal, angle), angle), 2))
-	ld velocity_soccer_kick(ld toGoal, ld height_at_goal, ld angle)const;
-
-	// targetDistance * sqrt((abs(acceleration)/((2 * (targetDistance * (tan(angle)*DEGREE) - targetHeight)))))
-	ld horizontal_velocity_using_distance_angle_height(ld targetDistance, ld targetHeight, ld angle, ld acceleration = GA)const;
-
-	// xVelocity * tan(angle*RADIAN)
-	ld vertical_velocity_by_Xvelocity_with_angle(ld xVelocity, ld angle)const;
-
-	//  atan(-((b)+sqrt((b * b) - 4 * a * c)) / (2 * a))*DEGREE;
-	//  atan(-((b)-sqrt((b * b) - 4 * a * c)) / (2 * a)) * DEGREE;
-	std::vector<ld> basketball_angles(ld launchVelocity, ld releaseHeight, ld hoopDistance)const;
-
-
-
-
 
 	/**========================================================================
 	 * overloaded operators
