+++ Physics_Formula_Class.vcxproj.filters
@@ -90,6 +90,9 @@
     <ClInclude Include="ElectricCurrent.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="Kinematics.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Vector2d.cpp">
